refactor(fs): unify file/directory implementations and enhance flags
- Unified 3 duplicate `File` structs and 3 `Directory` structs into single implementations
- Replaced legacy `OpenOptions` with extensible `OpenFlags` supporting full flag bits
- Fixed shell startup bug when running `make A=apps/fs/shell ARCH=aarch64 LOG=debug BLK=y run FEATURES=fatfs`

+++ modules/ruxhal/src/mem.rs
@@ -153,7 +153,7 @@ pub(crate) fn default_mmio_regions() -> impl Iterator<Item = MemRegion> {
 /// Returns the default MMIO memory regions (from [`ruxconfig::DTB_ADDR`]).
 #[allow(dead_code)]
 pub(crate) fn default_dtb_regions() -> impl Iterator<Item = MemRegion> {
-    let dtb_addr = PhysAddr::from(ruxconfig::DTB_ADDR as usize).align_up_4k();
+    let dtb_addr = PhysAddr::from(ruxconfig::DTB_ADDR).align_up_4k();
     let dtb_resgion = MemRegion {
         paddr: dtb_addr,
         size: 0x80000,
